requirements.txt
Flask==3.0.0
python-telegram-bot==13.15
Selenium==4.16.0
redis==5.0.1
SQLAlchemy==2.0.25
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
aiohttp==3.9.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
flask-cors==4.0.0
